src/requirements.txt
pysdl2
numpy
